Improved code quality
Solved linting issues in code. Including changing all the JSX-files to
*.jsx from *.js
