+++ src/pages/about.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import theme from '../styles/theme';
 
 import Layout from '../components/layout';
-import SEO from '../components/seo';
+import SEO from '../components/Seo';
 import Image from '../components/image';
 import Container from '../components/Container';
 import Header from '../components/Header.jsx';
