+++ src/pages/index.jsx
@@ -1,14 +1,10 @@
 import React from 'react';
-import { Link } from 'gatsby';
 // import P5Wrapper from 'react-p5-wrapper';
 
 import Layout from '../components/layout';
-import Image from '../components/image';
-import SEO from '../components/seo';
+import SEO from '../components/Seo';
 // import Sketch1 from "../components/sketch/Sketch.1";
 // import Sketch from "../components/sketch/Sketch";
-
-const testImg = require('../images/test-img.jpg');
 
 const IndexPage = () => (
   <Layout>
